pyproject.toml
[project]
name = "mcp-aws-dev"
version = "0.1.0"
description = "AWS Developer's Tools"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Martin Macak" }]
maintainers = [
    { name = "Martin Macak", email = "martin.macak@gmail.com" },
]
keywords = ["mcp", "llm", "automation", "aws"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "anyio>=4.5",
    "click>=8.1.0",
    "httpx>=0.27",
    "mcp",
    "boto3>=1.28.0",
    "typer>=0.15.2",
    "tomli>=2.2.1,<3.0.0",
    "pyyaml>=6.0.2",
    "jq (>=1.8.0,<2.0.0)",
    "docker>=7.1.0",
    "datamodel-code-generator>=0.30.0",
    "pydantic>=2.11.3",
    "requests-auth-aws-sigv4>=0.7",
    "requests>=2.32.3",
    "genson>=1.3.0",
]

[project.scripts]
mcp-aws-dev = "mcp_aws_dev.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_aws_dev"]

[tool.pyright]
include = ["mcp_aws_dev"]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.378",
    "pytest>=8.3.3",
    "ruff>=0.6.9",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "moto>=5.1.4",
    "mypy>=1.15.0",
    "pytest-mock>=3.14.0",
    "types-boto3>=1.37.38",
    "types-docker>=7.1.0.20250416",
]

[[tool.mypy.overrides]]
module = ["genson.*"]
follow_untyped_imports = true
